Run build graph successfully: - examples metadata and transcode - asyncify Demuxer custom io reader/seeker - enrich inferred format info - bigInt not work when enable asyncify, so binding int64_t to int - add `format` filter - define SourceConfig, FilterConfig, TargetConfig, with UserGraphNode - UserGraphNodes have source data, but no id. - Each export will generate a configNodes graph with ids - remove probeSize - add worker send to main thread with reply for message module - streamIO -> Reader can be seeked, create stream each time - Exporter.next do not need to provide input data - Worker -> InputStreamer handle retrieve data and used by C++ Demuxer

+++ src/cpp/metadata.h
@@ -1,7 +1,7 @@
 #ifndef METADATA_H
 #define METADATA_H
 
-// #include <cstdio>
+#include <cstdio>
 #include <string>
 #include <vector>
 
@@ -17,9 +17,14 @@ using namespace std;
 struct StreamInfo {
     int index;
     AVRational time_base;
-    int64_t bit_rate;
-    int64_t start_time;
-    int64_t duration;
+    
+    // int64_t bit_rate;
+    // int64_t start_time;
+    // int64_t duration;
+    int bit_rate;
+    int start_time;
+    int duration;
+    
     string codec_type;
     string codec_name;
     string format;
@@ -42,8 +47,10 @@ void set_avcodec_context_from_streamInfo(StreamInfo& info, AVCodecContext* ctx);
 
 struct FormatInfo {
     std::string format_name;
-    int64_t bit_rate;
-    int64_t duration;
+    // int64_t bit_rate;
+    // int64_t duration;
+    int bit_rate;
+    int duration;
     vector<StreamInfo> streamInfos;
 };
